Added a ColorChange function
Takes an index, for the corresponding line whose color you want to chnage, and the color you want to change it to. Changes the color for that line only. Had to make a change to how the lines were being drawn, they cant all be using the same material, but it was a quick and easy fix.